association_benchmark/requirements.txt
decord.egg==info
einops==0.7.0
ftfy==6.2.0
ipdb==0.13.13
mmcv_full==1.7.1
numpy==1.26.4
omegaconf==2.3.0
opencv_python==4.9.0.80
pandas==1.5.3
Pillow==10.2.0
regex==2023.12.25
setuptools==60.2.0
termcolor==2.4.0
timm==0.6.12
torch==1.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.65.2
transformers==4.29.0
wandb==0.13.7
